updated pyproject.toml
